requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
telethon==1.43.0
python-dotenv==1.0.0
pydantic==2.5.0
